requirements.txt

opencv-python
scikit-image
pillow
torch
torchvision
ninja
lmdb
requests
timm
scikit-learn
scipy
blobfile>=1.0.5
tqdm
matplotlib
